Refactor: separated db functions
